Add macelenia, rename best_lenia to lenia
