+++ profiles/lenia/automaton.py
