+++ profiles/lenia/clipvip/clip_vip.py
