+++ profiles/lenia/clipvip/miniblocks.py
