+++ profiles/lenia/configs/test.json
@@ -1,0 +1,18 @@
+{
+    "name":"test",
+    "epochs":30,
+    "val_split":0.1,
+    "loss":"cross_entropy",
+    "clipvip_size":32,
+    "minihead":true,
+    "early_stopping_patience":10,
+    "simulator": {
+        "size": [1,200,200],
+        "num_channels": 3,
+        "run_length": 1000,
+        "frame_portion": 5,
+        "num_output_frames": 12
+    },
+    "device":"cuda",
+    "test_set_path":"out/best_lenia/benchmark"
+}
