+++ profiles/lenia/generator.py
