--- profiles/lenia/simulator/utils/__init__.py
@@ -1,1 +1,0 @@
-from .leniaparams import LeniaParams
