+++ profiles/lenia/utils/adjectives.pk
