+++ profiles/lenia/utils/finder_utils.py
