+++ profiles/lenia/utils/funcgen.py
