+++ profiles/lenia/utils/noise_gen.py
