+++ profiles/lenia/utils/nouns.pk
