+++ profiles/macelenia/configs/macebase.json
@@ -1,13 +1,16 @@
 {
-    "name":"test",
+    "name":"macebase",
     "epochs":30,
     "val_split":0.1,
     "loss":"cross_entropy",
     "clipvip_size":32,
     "minihead":true,
     "early_stopping_patience":10,
+    "generator": {
+        "gen_mode": {"default_gen":false, "k_arbi":true, "g_arbi":false}
+    },
     "simulator": {
-        "size": [6,200,200],
+        "size": [1,200,200],
         "num_channels": 3,
         "run_length": 1000,
         "frame_portion": 5,
